Local Build Process: Take out the bumps.
When I started to try and get this site going on my machine I discovered
that it's pretty hard to setup correctly. A few of the issues with
hosting on GitHub Pages and Jekyll collided leading to some funky
instructions for doing local dev: commenting out variables in the build
config and also in CSS files. This process was a bit fragile with
regards to accidentally deploying changes that could take down the
production site.

Now, the build process has changed by the introduction of some
automatic variables that setup the environment in place of the
developer.

 + If the changes are accidentally deployed, they will reflect the
proper production environment.
 + Building locally required appending a local config file to the Jekyll
process: `jekyll --config _config.yml, _config_local.yml --baseurl ''
--watch`
 + The appropriate base url for CSS purposes has been moved to a
template variable in `style.css`
 + There is a new template variable, `production`, available which holds
`true` for production builds and `false` for development builds, as
specified by the inclusion of the local config file.

The local config file does not replace `_config.yml`, but merely
redefines a few variables, so both are needed.

I have also updated `README.md` to reflect these changes. Along the way
there were a few updates, such as using the `github-pages` **gem**
instead of installing all the dependencies separately.

Additionally I have linked some **git** tutorials in case this is all
very unfamiliar to some contributors.