Remove the deprecated calendar folder
Previously, the calendar view was rendered from within its own
subdirectory `/calendar/index.html`. During a previous refactor, this
and other components were brought into the `/pages` folder in order to
simplify the code of the site and take advantage of the `permalink`
front matter. When this happened, the calendar folder was left behind
causing conflicts because both here and in the new page the permalinks
and slugs were identical.

Now, without this folder everything works as expected.

N.B. This only appeared to impact the local development build of the
site, or perhaps the `gh-pages` branch was simply behind `master`.