Generates optimized versions of image assets
This commit changes the way that image assets are handled in the final
site build. Previously, each developer was responsible for optimizing
all of the image assets himself or herself. Now, the image optimization
is handled automatically to relieve that burder and to encourage
consistency across the images.

Total image assets before this commit amounted to 2.8 MB, but after the
optimization only 1.3 MB. This will further shrink as the image assets
are rearranged to take advantage of this new automatic step.

A follow-up commit will add the ability to specify the final image size
in the image filename, but currently all images are copied as-is but
limited to have the largest dimension no greater than 1600 pixels.

The `source_assets` folder is the new place to save the original images
and will also serve as the place in the future for all other
uncompressed/unminified assets.

Optimization is handled through the Makefile and occurs automatically.

**Changes to Makefile**
 - To run the local dev server run `make` or `make run` (`run` is the
   default target so you can leave it off).
 - To build the static files for the site but not start the server, run
   `make build`
 - To clear out the optimized assets (maybe to test or change the
   optimization parameters) run `make clean`

See `README.md` for normal development instructions.