Autogenerate optimized images on build with size
In the previous commit (bbc7d7307185bd8ce4cae35a57cc2de37653818e), I
introduced a pipeline for optimizing the images on every build so that
we could retain the high-quality source images in a separate folder and
keep them after optimization. The reason for doing this is so that we
can make changes to how we use those images without losing the source.

In this commit, I have enhanced that build process so that we can
specify the final image dimensions for the optimized copies. This is
important because if we want to start with a high-resolution version of
an image, we need to specify what the final size ought to be.

Size specification happens by ending the name of a source image with a
dash and then optionally the width, optionally followed by 'x' and the
height constraints.

	- "image.jpg" --> use original size but limit longest edge to 1600
	  pixels
	- "image-400.jpg" --> scale to set width to 400px and maintain
	  aspect ratio
	- "image-x200.jpg" --> scale to set height to 200px and maintain
	  aspect ratio
	- "image-400x200.jpg" --> scale and force height and width, butchers
	  aspect ratio