Fix rendering issues left-over from calendar
In #41 we removed the duplicate calendar page, but this inadvertently
produced a rendering error on the home page - a big green block where it
should have been white. This error was the result of a failed page
generation.

First of all, there was an error in the `navbar.html` template which I
have removed, though this was unrelated to the render issue.

Secondly, and more importantly, I believe that we ran into this issue
because there were two `index.html` files with overlapping slugs. After
removing the `index.html` file in the root directory, the rendering
glitch is gone.