Use a variable for the image directory base in <img> tags
Previously we were referencing the full image directory after the
basename in all of our image tags:

```html
<img src="{{ site.basename }}/assets/images/my_image.jpg" />
```

Now I have declared a new variable in `_config.yml` to use to help
prevent bugs from creeping in if that path changes and to standardize
the location naming:

```html
<img src="{{ site.basename }}/{{ site.img_dir }}/my_image.jpg" />
```

As you can see, it's not actually less writing to use this, however it
is more resilient to sneaky bugs. I explored creating a variable within
`_config.yml` that was composed of the site basename, but this does not
appear possible.

There is a project on GitHub to allow [using variables inside
_config.yml](https://github.com/gjtorikian/jekyll-config-variables) but
since we are constrained by GitHub's extensions, I do not believe this
will help us.