Include JavaScript libraries to fix broken nav menu
In #39 we removed the JavaScript libraries pulled in by the site's
template because they appeared to add 100K to the download size without
providing any actual benefit.

Apparently Bootstrap was relying on the JavaScript to open the
"hamburger" navigation menu when it was collapsed and clicked.

This patch loads the JavaScript back into the website, adding the
overhead again in order to protect the functionality.

This lends more priority weight to #46 because it shows a fragile aspect
of this site: we don't know where the dependencies exist in the code and
don't have a good way to minimize the final static output.