Use CDN links for JS instead of local copies
Since we are adding around 100K to the download for this site, we can at
least mitigate that overhead by referencing CDN versions of the code in
hopes that the visitors will have a locally-cached version of that file.
If it comes from our domain, then there will be no possibility for them
to have cached the scripts from other sites that also use them.