Make projects a separate collection
Attempts to finally resolve #20

Previously, the projects were defined inside of a data collection in
`projects.yml` in a tight format. I found this difficult to work with
because of the way the content squashed together and how leaving a
single line for the project description made it feel more like a plain
text string instead of a rich markup field.

Now, projects become their own collection found under the `_projects/`
folder, one file per project. Each project has front-matter specifying
data related to the project, and the `description` has been moved below
the front matter into the Markdown area and accessible as `content`.

Moving from a list of items into a folder of pages broke the inherent
sort order of the list, so I have updated `index.html` and
`projects.html` correspondingly to enforce a new sort based on the date
given for each project. I pulled the dates from the `git` history on the
`project.yml` file.

To illustrate the new flexibility I have updated a couple of the
descriptions for the City SDK project and the National Day of Civic
Hacking project with some markup such as links, bold, and italic
formatting.

Additionally I have removed the **Tucson Answers** project because I
could not find any reference to Tucson Answers with that search query on
Google other than this site, but the [GitHub
repository](https://github.com/CodeForTucson/tucson-answers) for it is
basically empty. Not really a flagship project, but we can pull it back
if we want.

P.S. Please pardon the `udpate` in the branch name. My bad.

cc: @meiqimichelle