Add missing project files
In #66 we pushed the latest batch of changes out to the production
`github-pages` branch, but all the projects disappeared. This was not a
problem with the collections interface, but a problem between the
keyboard and chair.

In a dicey `git` messup, I left the entire new `_projects` directory out
of the commit and thus although the project collection existed in
production, it had no data members.

Those who were responsible for this mishap have been sacked.