Setup Google Analytics and tracking
Resolves #3

Adds the appropriate snippet to load Google tracking services in the
website configured to the Code for Tucson Google account.

Uses the [Google Tag Manager](https://tagmanager.google.com/) to
administer the creation, modification, and removal of _tags_ such as
Google Analytics. Props to @dtwen for the suggestion against directly
embedding the analytics code and instead to use the tag manager.

Adds a basic page documenting the accounts and administrative process
for the analytics information.