Creates new focus pages
Related to:
 - #65 - Visual design reboot
 - #63 - Content reboot

Addresses desire to emphasize Communityshare and our educational
activities. By creating the two new pages in this PR, we should be able
to provide that empasis by giving first-class content and
discoverability to these projects and also by enumerating ways people
can jump in and contribute to them.