Emphasize Communityshare on homepage project list
Previously, we were simply grabbing the three most-recent projects
listed by their project date and displaying them on the home page.

This patch makes sure that the Communityshare project is displayed in
that list and also that it appears first. In order to support this
effort, I have refactored the HTML of the project row into a separate
`_include` file to try and make the logic of the display easier.